+++ logprep/util/pseudo/commands/pseudonymize.py
@@ -1,17 +1,28 @@
+"""Pseudonymize a string using the given keys and method."""
+
 import click
 
-from logprep.processor.pseudonymizer.encrypter import DualPKCS1HybridEncrypter
+from logprep.util.pseudo.encrypter import (
+    DualPKCS1HybridCTREncrypter,
+    DualPKCS1HybridGCMEncrypter,
+)
 
 
 @click.command()
-@click.argument("analyst-key", type=str)
-@click.argument("depseudo-key", type=str)
+@click.argument("analyst-key", type=click.Path(exists=True))
+@click.argument("depseudo-key", type=click.Path(exists=True))
 @click.argument("string", type=str)
-def pseudonymize(analyst_key: str, depseudo_key: str, string: str):
+@click.option(
+    "--mode",
+    type=click.Choice(["gcm", "ctr"]),
+    default="ctr",
+    help="The mode to use for decryption",
+)
+def pseudonymize(analyst_key: str, depseudo_key: str, string: str, mode: str):
     """pseudonymize a string using the given keys."""
-    encrypter = DualPKCS1HybridEncrypter()
+    encrypter = DualPKCS1HybridGCMEncrypter() if mode == "gcm" else DualPKCS1HybridCTREncrypter()
     encrypter.load_public_keys(
-        keyfile_analyst=f"{analyst_key}.crt",
-        keyfile_depseudo=f"{depseudo_key}.crt",
+        keyfile_analyst=f"{analyst_key}",
+        keyfile_depseudo=f"{depseudo_key}",
     )
     print(encrypter.encrypt(string))